Refactor controller spec: Separate controller runtime and customized reconcile spec
