+++ src/temporal-logic-lib/examples/kubernetes_cluster/proof/controller_runtime_liveness.rs
@@ -5,8 +5,13 @@ use crate::examples::kubernetes_cluster::{
     proof::{kubernetes_api_safety, wf1_assistant::controller_action_pre_implies_next_pre},
     spec::{
         common::*,
-        controller,
-        controller::{ending_step, relevant_cr_key, ControllerAction, ControllerActionInput},
+        controller::common::{
+            ending_step, ControllerAction, ControllerActionInput, ReconcileCoreStep,
+        },
+        controller::controller_runtime::{
+            continue_reconcile, end_reconcile, run_scheduled_reconcile, trigger_reconcile,
+        },
+        controller::state_machine::controller,
         distributed_system::*,
     },
 };
@@ -20,7 +25,7 @@ verus! {
 
 pub proof fn lemma_pre_leads_to_post_by_controller(input: ControllerActionInput, action: ControllerAction, pre: StatePred<State>, post: StatePred<State>)
     requires
-        controller::controller().actions.contains(action),
+        controller().actions.contains(action),
         forall |s, s_prime: State| pre(s) && #[trigger] next()(s, s_prime) ==> pre(s_prime) || post(s_prime),
         forall |s, s_prime: State| pre(s) && #[trigger] next()(s, s_prime) && controller_next().forward(input)(s, s_prime) ==> post(s_prime),
         forall |s: State| #[trigger] pre(s) ==> controller_action_pre(action, input)(s),
@@ -37,7 +42,7 @@ pub proof fn lemma_pre_leads_to_post_by_controller(input: ControllerActionInput,
 
 pub proof fn lemma_pre_leads_to_post_with_asm_by_controller(input: ControllerActionInput, action: ControllerAction, asm: StatePred<State>, pre: StatePred<State>, post: StatePred<State>)
     requires
-        controller::controller().actions.contains(action),
+        controller().actions.contains(action),
         forall |s, s_prime: State| pre(s) && #[trigger] next()(s, s_prime) && asm(s) ==> pre(s_prime) || post(s_prime),
         forall |s, s_prime: State| pre(s) && #[trigger] next()(s, s_prime) && controller_next().forward(input)(s, s_prime) ==> post(s_prime),
         forall |s: State| #[trigger] pre(s) ==> controller_action_pre(action, input)(s),
@@ -50,44 +55,6 @@ pub proof fn lemma_pre_leads_to_post_with_asm_by_controller(input: ControllerAct
     valid_implies_trans::<State>(lift_state(pre), lift_state(controller_action_pre(action, input)), lift_state(controller_next().pre(input)));
 
     controller_next().wf1_assume(input, sm_spec(), next(), asm, pre, post);
-}
-
-pub proof fn lemma_relevant_event_sent_leads_to_reconcile_triggered(msg: Message, cr_key: ResourceKey)
-    requires
-        cr_key.kind.is_CustomResourceKind(),
-    ensures
-        sm_spec().entails(
-            lift_state(|s: State| {
-                &&& s.message_sent(msg)
-                &&& msg.dst === HostId::CustomController
-                &&& msg.content.is_WatchEvent()
-                &&& relevant_cr_key(msg) === Option::Some(cr_key)
-                &&& !s.reconcile_state_contains(cr_key)
-            })
-                .leads_to(lift_state(|s: State| {
-                    &&& s.reconcile_state_contains(cr_key)
-                    &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
-                    &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
-                }))
-        ),
-{
-    let pre = |s: State| {
-        &&& s.message_sent(msg)
-        &&& msg.dst === HostId::CustomController
-        &&& msg.content.is_WatchEvent()
-        &&& relevant_cr_key(msg) === Option::Some(cr_key)
-        &&& !s.reconcile_state_contains(cr_key)
-    };
-    let post = |s: State| {
-        &&& s.reconcile_state_contains(cr_key)
-        &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
-        &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
-    };
-    let input = ControllerActionInput {
-        recv: Option::Some(msg),
-        scheduled_cr_key: Option::None,
-    };
-    lemma_pre_leads_to_post_by_controller(input, controller::trigger_reconcile(), pre, post);
 }
 
 pub proof fn lemma_reconcile_ended_leads_to_reconcile_ended(cr_key: ResourceKey)
@@ -117,7 +84,7 @@ pub proof fn lemma_reconcile_ended_leads_to_reconcile_ended(cr_key: ResourceKey)
         recv: Option::None,
         scheduled_cr_key: Option::Some(cr_key),
     };
-    lemma_pre_leads_to_post_by_controller(input, controller::end_reconcile(), pre, post);
+    lemma_pre_leads_to_post_by_controller(input, end_reconcile(), pre, post);
 }
 
 pub proof fn lemma_scheduled_reconcile_leads_to_init(cr_key: ResourceKey)
@@ -131,7 +98,7 @@ pub proof fn lemma_scheduled_reconcile_leads_to_init(cr_key: ResourceKey)
             })
                 .leads_to(lift_state(|s: State| {
                     &&& s.reconcile_state_contains(cr_key)
-                    &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
+                    &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::Init
                     &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
                 }))
         ),
@@ -142,14 +109,14 @@ pub proof fn lemma_scheduled_reconcile_leads_to_init(cr_key: ResourceKey)
     };
     let post = |s: State| {
         &&& s.reconcile_state_contains(cr_key)
-        &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
+        &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::Init
         &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
     };
     let input = ControllerActionInput {
         recv: Option::None,
         scheduled_cr_key: Option::Some(cr_key),
     };
-    lemma_pre_leads_to_post_by_controller(input, controller::run_scheduled_reconcile(), pre, post);
+    lemma_pre_leads_to_post_by_controller(input, run_scheduled_reconcile(), pre, post);
 }
 
 pub proof fn lemma_reconcile_ended_leads_to_init(cr_key: ResourceKey)
@@ -163,7 +130,7 @@ pub proof fn lemma_reconcile_ended_leads_to_init(cr_key: ResourceKey)
             })
                 .leads_to(lift_state(|s: State| {
                     &&& s.reconcile_state_contains(cr_key)
-                    &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
+                    &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::Init
                     &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
                 }))
         ),
@@ -180,7 +147,7 @@ pub proof fn lemma_reconcile_ended_leads_to_init(cr_key: ResourceKey)
     };
     let reconcile_at_init = |s: State| {
         &&& s.reconcile_state_contains(cr_key)
-        &&& s.reconcile_state_of(cr_key).reconcile_step === controller::ReconcileCoreStep::Init
+        &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::Init
         &&& s.reconcile_state_of(cr_key).pending_req_msg.is_None()
     };
     leads_to_trans::<State>(sm_spec(), reconcile_ended, reconcile_rescheduled, reconcile_at_init);
