+++ src/temporal-logic-lib/examples/kubernetes_cluster/proof/controller_runtime_safety.rs
@@ -6,7 +6,7 @@ use crate::examples::kubernetes_cluster::{
     spec::{
         common::*,
         controller,
-        controller::{relevant_cr_key, ControllerAction, ControllerActionInput, ReconcileCoreStep},
+        controller::common::{ControllerAction, ControllerActionInput, ReconcileCoreStep},
         distributed_system::*,
     },
 };
