+++ src/temporal-logic-lib/examples/kubernetes_cluster/proof/reconcile_safety.rs
@@ -5,8 +5,8 @@ use crate::examples::kubernetes_cluster::{
     proof::kubernetes_api_safety,
     spec::{
         common::*,
-        controller,
-        controller::{relevant_cr_key, ControllerAction, ControllerActionInput, ReconcileCoreStep},
+        controller::common::{ControllerAction, ControllerActionInput, ReconcileCoreStep},
+        controller::simple_controller,
         distributed_system::*,
     },
 };
@@ -71,8 +71,8 @@ pub proof fn lemma_always_reconcile_create_cm_done_implies_pending_create_cm_req
                 .implies(lift_state(|s: State| {
                     &&& s.reconcile_state_contains(cr_key)
                     &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::CreateCMDone
-                    &&& s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
-                    &&& s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
+                    &&& s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
+                    &&& s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
                 }))
         )),
 {
@@ -81,8 +81,8 @@ pub proof fn lemma_always_reconcile_create_cm_done_implies_pending_create_cm_req
         && s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::CreateCMDone
         ==> s.reconcile_state_contains(cr_key)
             && s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::CreateCMDone
-            && s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
-            && s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
+            && s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
+            && s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
     };
     init_invariant::<State>(sm_spec(), init(), next(), invariant);
 
@@ -92,8 +92,8 @@ pub proof fn lemma_always_reconcile_create_cm_done_implies_pending_create_cm_req
     }).implies(lift_state(|s: State| {
         &&& s.reconcile_state_contains(cr_key)
         &&& s.reconcile_state_of(cr_key).reconcile_step === ReconcileCoreStep::CreateCMDone
-        &&& s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
-        &&& s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(controller::create_cm_req(cr_key))))
+        &&& s.reconcile_state_of(cr_key).pending_req_msg === Option::Some(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
+        &&& s.message_sent(form_msg(HostId::CustomController, HostId::KubernetesAPI, MessageContent::APIRequest(simple_controller::create_cm_req(cr_key))))
     }));
     temp_pred_equality::<State>(lift_state(invariant), invariant_temp_pred);
 }
